// src/gen/api.zod.ts
export const addressControllerCreateBody = zod.object({
  "line_1": zod.string(),
  "line_2": zod.string().optional(),
  "zip_code": zod.string(),
  "city": zod.string(),
  "state": zod.string(),
  "country": zod.string().optional(),
  "label": zod.string().optional(),
  "is_default": zod.boolean().optional()
})